Add existing code into the correct workspace
