+++ .gitignore
@@ -1,1 +1,4 @@
 /target
+/data
+!/data/valid.json
+!/data/invalid.json
