+++ tracker/lib/src/lib.rs
@@ -1,3 +1,9 @@
+#[macro_use]
+extern crate lazy_static;
+
+mod assignment;
+pub use assignment::Assignment;
+
 #[cfg(test)]
 mod tests {
     #[test]
